fix servers errors
