+++ .env.example
@@ -6,3 +6,9 @@ UKRNET_AUTH_USER= # ukr.net IMAP auth email
 UKRNET_AUTH_PASS= # ukr.net IMAP auth token
 UKRNET_SENDER= # ukr.net sender email
 SUPPORT_EMAIL= # Support email
+MJ_APIKEY_PUBLIC= #
+MJ_APIKEY_PRIVATE= #
+CLOUDINARY_URL= #
+CLOUDINARY_KEY= #
+CLOUDINARY_SECRET= #
+CLOUDINARY_CLOUD_NAME= #
