+++ controllers/userControllers/sendHelpMsg.js
@@ -3,14 +3,27 @@ import sendEmail from "../../helpers/sendEmail.js";
 export async function sendHelpMsg(req, res, next) {
   const { email, comment } = req.body;
 
-  const helpEmail = {
-    to: process.env.SUPPORT_EMAIL,
-    from: process.env.UKRNET_SENDER,
-    subject: "Need Help",
-    html: `
-      <div>${comment}</div></br>
-      <div><a href="mailto:${email}" target="_blank">${email}</a></div>`,
-    text: `${comment} ${email}`,
+  const helpEmail = {    Messages: [
+      {
+        From: {
+          Email: process.env.UKRNET_SENDER,
+        },
+        To: [
+          {
+            Email: process.env.SUPPORT_EMAIL,
+          },
+          {
+            Email: process.env.SUB_SUPPORT_EMAIL,
+          },
+        ],
+        Subject: "Need help",
+        HTMLPart: `
+            <div>${comment}</div></br>
+            <div><a href="mailto:${email}" target="_blank">${email}</a></div>`,
+        TEXTPart: `${comment} ${email}`,
+        CustomID: "AppGettingStartedTest",
+      },
+    ],
   };
 
   await sendEmail(helpEmail);
