+++ package.json
@@ -18,6 +18,7 @@
     "mongoose": "^8.3.3",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
+    "node-mailjet": "^6.0.5",
     "nodemailer": "^6.9.13",
     "swagger-ui-express": "^5.0.1"
   },
